code/backend/twitter/requirements_rest.txt
asgiref==3.2.5
attrs==19.3.0
bson==0.5.8
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.3
dataclasses==0.6
Django==2.2.11
django-cors-headers==3.2.1
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djongo==1.3.0
Faker==0.9.1
idna==2.9
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mixer==6.1.3
more-itertools==8.2.0
neo4j-driver==1.7.6
neobolt==1.7.16
neotime==1.7.4
openapi-codec==1.3.2
packaging==20.3
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.8.1
pymongo==3.10.1
pyparsing==2.4.6
pytest==5.3.5
pytest-cov==2.8.1
pytest-django==3.8.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.2.4
text-unidecode==1.2
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.8
django-redis==4.11.0
redis==3.5.3
